Improve Rainbow mode
Previously, the app made use of Math.random to generate a random
background color for each "hovered" grid in the container.

Instead of randomly generating the color, the app now uses client's
viewport width and height and their mouse's position to give the grid
darker/lighter shades of red, green and blue colors.